Make `ConfigRepository` wrapper unconditional
