+++ core/lib/config/src/lib.rs
@@ -1,20 +1,22 @@
 #![allow(clippy::upper_case_acronyms, clippy::derive_partial_eq_without_eq)]
 
-pub use crate::configs::{
-    contracts::chain::ContractsConfig, full_config_schema, ApiConfig, AvailConfig,
-    BaseTokenAdjusterConfig, CelestiaConfig, ContractVerifierConfig, DAClientConfig,
-    DADispatcherConfig, DBConfig, EigenConfig, EthConfig, EthWatchConfig,
-    ExternalProofIntegrationApiConfig, GasAdjusterConfig, GenesisConfig, ObjectStoreConfig,
-    PostgresConfig, SnapshotsCreatorConfig,
+pub use crate::{
+    configs::{
+        contracts::chain::ContractsConfig, full_config_schema, ApiConfig, AvailConfig,
+        BaseTokenAdjusterConfig, CelestiaConfig, ContractVerifierConfig, DAClientConfig,
+        DADispatcherConfig, DBConfig, EigenConfig, EthConfig, EthWatchConfig,
+        ExternalProofIntegrationApiConfig, GasAdjusterConfig, GenesisConfig, ObjectStoreConfig,
+        PostgresConfig, SnapshotsCreatorConfig,
+    },
+    repository::{CapturedParams, ConfigRepository},
 };
-#[cfg(feature = "observability_ext")]
-pub use crate::observability_ext::{CapturedParams, ConfigRepository};
 
 #[cfg(feature = "cli")]
 pub mod cli;
 pub mod configs;
 #[cfg(feature = "observability_ext")]
 mod observability_ext;
+mod repository;
 pub mod sources;
 #[cfg(test)]
 mod tests;
