+++ core/lib/config/src/sources.rs
@@ -4,7 +4,9 @@ use std::{
 };
 
 use anyhow::Context;
-use smart_config::{Environment, Prefixed, Yaml};
+use smart_config::{ConfigSchema, DescribeConfig, Environment, Prefixed, Yaml};
+
+use crate::{configs::ObservabilityConfig, repository::log_all_errors, ConfigRepository};
 
 /// Wrapper around configuration sources.
 #[derive(Debug, Default)]
@@ -17,15 +19,27 @@ impl ConfigSources {
         Ok(self)
     }
 
-    #[cfg(any(test, feature = "observability_ext"))]
-    pub(crate) fn build_raw_repository(
-        self,
-        schema: &smart_config::ConfigSchema,
-    ) -> smart_config::ConfigRepository<'_> {
+    fn build_raw_repository(self, schema: &ConfigSchema) -> smart_config::ConfigRepository<'_> {
         let mut repo = smart_config::ConfigRepository::new(schema);
         repo.deserializer_options().coerce_variant_names = true;
         repo.deserializer_options().coerce_serde_enums = true;
         repo.with_all(self.0)
+    }
+
+    /// Returns the observability config. It should be used to install observability early in the executable lifecycle.
+    pub fn observability(&self) -> anyhow::Result<ObservabilityConfig> {
+        let schema = ConfigSchema::new(&ObservabilityConfig::DESCRIPTION, "observability");
+        let mut repo = smart_config::ConfigRepository::new(&schema).with_all(self.0.clone());
+        repo.deserializer_options().coerce_variant_names = true;
+        // - `unwrap()` is safe: `Self` is the only top-level config, so an error would require for it to have a recursive definition.
+        // - While logging is not enabled at this point, we use `log_all_errors()` for more intelligent error summarization.
+        repo.single().unwrap().parse().map_err(log_all_errors)
+    }
+
+    /// Builds the repository with the specified config schema. Deserialization options are tuned to be backward-compatible
+    /// with the existing file-based configs (e.g., coerce enum variant names).
+    pub fn build_repository(self, schema: &ConfigSchema) -> ConfigRepository<'_> {
+        self.build_raw_repository(schema).into()
     }
 }
